Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
supabase==2.5.0
gotrue==2.5.0
postgrest==0.15.0
python-dotenv==1.0.0
email-validator==2.1.0
PyJWT==2.12.1
reportlab==4.4.10
